Move some logout functions out of saml.ts
